Fixed compile warnings

+++ STM32/Util/Src/systeminfo.c
@@ -41,9 +41,9 @@ static char buf[600] = { 0 };
     #define ID2 *((unsigned long *) (UID_BASE + 4U))
     #define ID3 *((unsigned long *) (UID_BASE + 8U))
 #else
-    #define ID1 0U
-    #define ID2 0U
-    #define ID3 0U
+    #define ID1 (unsigned long) 0U
+    #define ID2 (unsigned long) 0U
+    #define ID3 (unsigned long) 0U
 
     #define GIT_VERSION "0"
     #define GIT_DATE    "0"
